+++ canon-docs/src/components/PropsTable/PropsTable.tsx
@@ -6,8 +6,9 @@ import { icons } from '../../../../packages/canon';
 
 // Define a more specific type for the data object
 type PropData = {
-  type: string | string[];
-  responsive: boolean;
+  values?: string | string[];
+  responsive?: boolean;
+  default?: string;
 };
 
 // Modify the PropsTable component to use the new type
@@ -16,17 +17,34 @@ export const PropsTable = <T extends Record<string, PropData>>({
 }: {
   data: T;
 }) => {
+  const completeData = {
+    ...data,
+    children: {
+      values: 'ReactNode',
+      responsive: false,
+    },
+    className: {
+      values: 'string',
+      responsive: false,
+    },
+    style: {
+      values: 'CSSProperties',
+      responsive: false,
+    },
+  };
+
   return (
     <Table.Root>
       <Table.Header>
         <Table.HeaderRow>
           <Table.HeaderCell>Prop</Table.HeaderCell>
           <Table.HeaderCell>Type</Table.HeaderCell>
+          <Table.HeaderCell>Default</Table.HeaderCell>
           <Table.HeaderCell>Responsive</Table.HeaderCell>
         </Table.HeaderRow>
       </Table.Header>
       <Table.Body>
-        {Object.keys(data).map(n => (
+        {Object.keys(completeData).map(n => (
           <Table.Row key={n}>
             <Table.Cell>
               <Chip head>{n}</Chip>
@@ -35,17 +53,22 @@ export const PropsTable = <T extends Record<string, PropData>>({
               <div
                 style={{ display: 'flex', flexWrap: 'wrap', gap: '0.375rem' }}
               >
-                {data[n].type === 'icon' ? (
+                {completeData[n].values === 'icon' ? (
                   Object.keys(icons).map(icon => <Chip key={icon}>{icon}</Chip>)
-                ) : Array.isArray(data[n].type) ? (
-                  data[n].type.map(t => <Chip key={t}>{t}</Chip>)
+                ) : Array.isArray(completeData[n].values) ? (
+                  completeData[n].values.map(t => <Chip key={t}>{t}</Chip>)
                 ) : (
-                  <Chip>{data[n].type}</Chip>
+                  <Chip>{completeData[n].values}</Chip>
                 )}
               </div>
             </Table.Cell>
             <Table.Cell>
-              <Chip>{data[n].responsive ? 'Yes' : 'No'}</Chip>
+              <Chip>
+                {completeData[n].default ? completeData[n].default : '-'}
+              </Chip>
+            </Table.Cell>
+            <Table.Cell>
+              <Chip>{completeData[n].responsive ? 'Yes' : 'No'}</Chip>
             </Table.Cell>
           </Table.Row>
         ))}