+++ plugins/events-node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-events-node
 
+## 0.4.8-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.2.0-next.2
+  - @backstage/errors@1.2.7
+  - @backstage/types@1.2.1
+
 ## 0.4.8-next.1
 
 ### Patch Changes